# web/public/openapi.yaml
openapi: 3.0.3
info:
  title: Pemuda Surau Al-Abqori API
  version: 1.0.0
  description: |
    REST API for the Surau Al-Abqori youth programme: prayer attendance
    (face-verified + geofenced), Quran activity, merits, programs and the
    public dashboard.

    ## Authentication
    Most endpoints require a JWT obtained from `POST /api/auth/verify-otp`.
    Send it as `Authorization: Bearer <token>`.

    ## Authorisation
    Beyond "logged in", endpoints are gated by **capabilities** rather than raw
    roles (see `server/src/utils/roles.js`). Admins implicitly hold every
    capability. The `x-capability` extension on each operation records the
    requirement.

    ## Geofencing
    Check-in, activity recording and new registrations are only accepted within
    a radius of the surau. Coordinates are supplied by the client.
servers:
  - url: /api
    description: Same origin (proxied by Nginx / the Vite dev server)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Authentication required
        code:
          type: string
          description: Optional machine-readable code (e.g. REGISTRATION_TOO_FAR).
          example: REGISTRATION_TOO_FAR
        retryAfterMinutes:
          type: integer
          description: Present on 429 responses (e.g. the Quran log cooldown).
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        phone: { type: string, nullable: true, example: '+60123456789' }
        fullName: { type: string, example: Ahmad bin Zainal }
        birthDate: { type: string, format: date, nullable: true }
        age: { type: integer, nullable: true, example: 17 }
        gender: { type: string, nullable: true, enum: [male, female] }
        address: { type: string, nullable: true }
        role: { type: string, enum: [youth, admin, parent, teacher, ajk] }
        capabilities:
          type: array
          items: { type: string }
          example: [checkIn, viewMembers]
        hasFace: { type: boolean }
        telegramLinked: { type: boolean }
        guardianId: { type: string, format: uuid, nullable: true }
        isDependent: { type: boolean }
        canSelfCheckIn: { type: boolean }
        createdAt: { type: string, format: date-time }
    Descriptor:
      type: array
      description: 128-float face descriptor produced by face-api.js in the browser.
      items: { type: number }
      minItems: 128
      maxItems: 128
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Invalid request payload.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Authenticated but lacking the required capability.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
paths:
...
  /users/me:
    put:
      tags: [Users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName]
              properties:
                fullName: { type: string }
                birthDate: { type: string, format: date, nullable: true }
                gender: { type: string, enum: [male, female], nullable: true }
                address: { type: string, nullable: true }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /users/me/face:
    post:
      tags: [Users]
      summary: Enroll (or re-enroll) own face
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [descriptor]
              properties:
                descriptor: { $ref: '#/components/schemas/Descriptor' }
      responses:
        '200':
          description: Enrolled
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /users/me/dependents:
    get:
      tags: [Users]
      summary: List my dependents (children)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  dependents:
                    type: array
                    items: { $ref: '#/components/schemas/User' }
    post:
      tags: [Users]
      summary: Add a dependent
      description: Dependents have no phone and cannot log in until given one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName]
              properties:
                fullName: { type: string }
                birthDate: { type: string, format: date, nullable: true }
                gender: { type: string, enum: [male, female], nullable: true }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  dependent: { $ref: '#/components/schemas/User' }
  /users/me/dependents/{id}:
    put:
      tags: [Users]
      summary: Update a dependent
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fullName]
              properties:
                fullName: { type: string }
                birthDate: { type: string, format: date, nullable: true }
                gender: { type: string, enum: [male, female], nullable: true }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  dependent: { $ref: '#/components/schemas/User' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Users]
      summary: Remove a dependent
      description: Cascades to their attendance, Quran logs and merits.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200': { description: Removed }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/me/dependents/{id}/face:
    post:
      tags: [Users]
      summary: Enroll a dependent's face
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [descriptor]
              properties:
                descriptor: { $ref: '#/components/schemas/Descriptor' }
      responses:
        '200': { description: Enrolled }
        '400': { $ref: '#/components/responses/BadRequest' }
  /users/me/dependents/{id}/phone:
    post:
      tags: [Users]
      summary: Give a dependent their own login
      description: |
        Sets a phone number on the existing dependent row so they can log in and
        check themselves in. History and the enrolled face are preserved, and
        the guardian link is kept. Guardians may do this for their own
        dependents; admins may do it for any dependent.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, example: '0123456789' }
      responses:
        '200':
          description: Promoted
          content:
            application/json:
              schema:
                type: object
                properties:
                  dependent: { $ref: '#/components/schemas/User' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: That phone number is already registered to another account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /users/me/dependents/{id}/co-guardian:
    post:
      tags: [Users]
      summary: Share a dependent with a second guardian
      description: |
        Sets (or clears) a dependent's second guardian so both parents can check
        the child in — they often arrive separately. The other parent must
        already have an account; they are looked up by phone number. Only the
        primary guardian (or an admin) may change this. Send `phone: null` to
        clear the co-guardian.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  nullable: true
                  example: '0123456789'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  dependent: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users:
    get:
      tags: [Users]
      summary: List all members
      description: Requires the `viewMembers` capability (teachers/AJK need it to pick a member).
      x-capability: viewMembers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      allOf:
                        - $ref: '#/components/schemas/User'
                        - type: object
                          properties:
                            guardianName: { type: string, nullable: true }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/{id}/role:
    patch:
      tags: [Users]
      summary: Change a member's role
      description: Admin only. Dependents inherit their guardian's role and cannot be changed.
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [youth, admin, parent, teacher, ajk] }
      responses:
        '200': { description: Updated }
        '400': { $ref: '#/components/responses/BadRequest' }
  /users/{id}/active:
    patch:
      tags: [Users]
      summary: Activate or deactivate a member
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              properties:
                isActive: { type: boolean }
      responses:
        '200': { description: Updated }
  /users/{id}/login-code:
    post:
      tags: [Users]
      summary: Generate an admin-assisted login code
      description: |
        Admin only. Last resort when a member cannot receive an OTP — the code
        is returned in the response so an admin can read it out.
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Code generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  code: { type: string, example: '123456' }
                  phone: { type: string }
                  fullName: { type: string }
                  expiresAt: { type: string, format: date-time }
        '400':
          description: Dependent accounts have no phone to log in with.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ------------------------------------------------------------ Attendance
...